Move test_fault_handler.h into test_runner.c
