+++ tests/destructors/main.c
@@ -6,7 +6,7 @@
 #include <ia2.h>
 #include "plugin.h"
 #define IA2_DEFINE_TEST_HANDLER
-#include "test_fault_handler.h"
+
 
 // This test uses two protection keys
 INIT_RUNTIME(2);
