+++ tests/destructors/plugin.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <ia2.h>
 #include "exported_fn.h"
-#include "test_fault_handler.h"
+
 
 #define IA2_COMPARTMENT 2
 #include <ia2_compartment_init.inc>
