+++ tests/mmap_loop/main.c
@@ -9,7 +9,7 @@ RUN: sh -c 'if [ ! -s "mmap_loop_call_gates_0.ld" ]; then echo "No link args as 
 #include <ia2_test_runner.h>
 
 #define IA2_DEFINE_TEST_HANDLER
-#include "test_fault_handler.h"
+
 
 /*
     This program tests that mmap and heap allocations are handled properly.
