+++ tests/protected_threads/main.c
@@ -9,7 +9,7 @@ RUN: sh -c 'if [ ! -s "protected_threads_call_gates_0.ld" ]; then echo "No link 
 #include <sys/wait.h>
 #define IA2_DEFINE_TEST_HANDLER
 #include <ia2_test_runner.h>
-#include <test_fault_handler.h>
+
 #include <unistd.h>
 
 INIT_RUNTIME(2);
