+++ tests/recursion/main.c
@@ -8,7 +8,7 @@ RUN: cat recursion_call_gates_2.ld | FileCheck --check-prefix=LINKARGS %s
 #include <ia2.h>
 #include <stdio.h>
 #define IA2_DEFINE_TEST_HANDLER
-#include "test_fault_handler.h"
+
 
 INIT_RUNTIME(2);
 #define IA2_COMPARTMENT 1
