+++ tests/ro_sharing/main.c
@@ -7,7 +7,7 @@ RUN: sh -c 'if [ ! -s "ro_sharing_call_gates_0.ld" ]; then echo "No link args as
 #include <ia2.h>
 #include <stdio.h>
 #define IA2_DEFINE_TEST_HANDLER
-#include "test_fault_handler.h"
+
 
 // This test checks that all RO data mapped in from executable files is shared.
 // This is needed so that the dynamic linker can read ELF metadata. Read-only
