+++ tests/ro_sharing/plugin.c
@@ -1,7 +1,7 @@
 /*
 RUN: cat ro_sharing_call_gates_1.ld | FileCheck --check-prefix=LINKARGS %s
 */
-#include "test_fault_handler.h"
+
 #include <ia2_test_runner.h>
 
 #include <ia2.h>
