+++ tests/should_segfault/main.c
@@ -8,7 +8,7 @@ RUN: sh -c 'if [ ! -s "should_segfault_call_gates_0.ld" ]; then echo "No link ar
 #include <print_secret.h>
 #include <ia2.h>
 #define IA2_DEFINE_TEST_HANDLER
-#include "test_fault_handler.h"
+
 
 INIT_RUNTIME(1);
 #define IA2_COMPARTMENT 1
