+++ tests/should_segfault/print_secret.c
@@ -6,7 +6,7 @@ RUN: cat should_segfault_call_gates_1.ld | FileCheck --check-prefix=LINKARGS %s
 #include <stdlib.h>
 #include <stdbool.h>
 #include "print_secret.h"
-#include "test_fault_handler.h"
+
 
 static bool early_fault = false;
