+++ tests/tls_protected/main.c
@@ -6,7 +6,7 @@
 #include <stdint.h>
 #include <unistd.h>
 #define IA2_DEFINE_TEST_HANDLER
-#include "test_fault_handler.h"
+
 #include <threads.h>
 
 INIT_RUNTIME(2);
