+++ tests/trusted_direct/main.c
@@ -10,7 +10,7 @@ RUN: cat trusted_direct_call_gates_0.ld | FileCheck --check-prefix=LINKARGS %s
 #include <ia2.h>
 #include "plugin.h"
 #define IA2_DEFINE_TEST_HANDLER
-#include "test_fault_handler.h"
+
 
 // This test checks that an untrusted library can call a trusted main binary. An
 // MPK violation is triggered from the untrusted library if no arguments are
