+++ tests/trusted_direct/plugin.c
@@ -4,7 +4,7 @@ RUN: cat trusted_direct_call_gates_1.ld | FileCheck --check-prefix=LINKARGS %s
 #include <ia2_test_runner.h>
 
 #include "exported_fn.h"
-#include "test_fault_handler.h"
+
 
 // LINKARGS: --wrap=start_plugin
 void start_plugin(void) {
