+++ tests/two_shared_ranges/main.c
@@ -10,7 +10,7 @@ RUN: cat two_shared_ranges_call_gates_2.ld | FileCheck --check-prefix=LINKARGS %
 #include <ia2.h>
 #include "plugin.h"
 #define IA2_DEFINE_TEST_HANDLER
-#include "test_fault_handler.h"
+
 
 // This test uses two protection keys
 INIT_RUNTIME(2);
