+++ tests/two_shared_ranges/plugin.c
@@ -8,7 +8,7 @@ RUN: cat two_shared_ranges_call_gates_1.ld | FileCheck --check-prefix=LINKARGS %
 #include <ia2_test_runner.h>
 #include <ia2.h>
 #include "exported_fn.h"
-#include "test_fault_handler.h"
+
 
 #define IA2_COMPARTMENT 2
 #include <ia2_compartment_init.inc>
