+++ tests/untrusted_indirect/main.c
@@ -7,7 +7,7 @@ RUN: sh -c 'if [ ! -s "untrusted_indirect_call_gates_0.ld" ]; then echo "No link
 #include "foo.h"
 #include <ia2.h>
 #define IA2_DEFINE_TEST_HANDLER
-#include "test_fault_handler.h"
+
 
 INIT_RUNTIME(1);
 #define IA2_COMPARTMENT 1
